+++ src/App.js
@@ -16,6 +16,7 @@ import PrivateRoute from './components/private-route';
 import NewMoviePage from './pages/newmovie';
 import MovieDetailsPage from './pages/movie-details';
 import BottomAppBar from './components/bottom-nav-bar';
+import ListPopularMovies from './components/list-popular-movies';
 
 function App() {
     const theme = createMuiTheme({
@@ -43,6 +44,7 @@ function App() {
                 <Container fixed>
                     <Switch>
                         <PrivateRoute path="/movies" component={ListMovies} />
+                        <Route path="/popular" component={ListPopularMovies} />
                         <Route path="/newmovie" component={NewMoviePage} />
                         <Route exact path="/movie-details/:id" component={MovieDetailsPage} />