+++ src/components/moviecard/index.js
@@ -1,31 +1,45 @@
 import React from 'react';
-import { makeStyles } from '@material-ui/core/styles';
+import { makeStyles, useTheme } from '@material-ui/core/styles';
 import Card from '@material-ui/core/Card';
-import { CardMedia, CardActions, CardContent } from '@material-ui/core';
-import Button from '@material-ui/core/Button';
+import CardContent from '@material-ui/core/CardContent';
+import CardMedia from '@material-ui/core/CardMedia';
+import IconButton from '@material-ui/core/IconButton';
 import Typography from '@material-ui/core/Typography';
+import FavoriteBorder from '@material-ui/icons/FavoriteBorder';
 import { useHistory } from 'react-router-dom';
 
-const useStyles = makeStyles({
+const useStyles = makeStyles((theme) => ({
     root: {
-        minWidth: 275,
+        display: 'flex',
+        width: 370,
         marginTop: 15,
-        width: '100%',
     },
-    media: {
-        height: 0,
-        paddingTop: '56.25%', // 16:9
+    details: {
+        display: 'flex',
+        flexDirection: 'column',
     },
-    title: {
-        fontSize: 14,
+    content: {
+        flex: '1 0 auto',
+        width: 266,
     },
-    pos: {
-        marginBottom: 12,
+    cover: {
+        width: 105,
     },
-});
+    controls: {
+        display: 'flex',
+        alignItems: 'center',
+        paddingLeft: theme.spacing(1),
+        paddingBottom: theme.spacing(1),
+    },
+    favIcon: {
+        height: 38,
+        width: 38,
+    },
+}));
 
 export default function MovieCard({ movie }) {
     const classes = useStyles();
+    const theme = useTheme();
     let history = useHistory();
 
     const openMovieDetails = () => {
@@ -34,29 +48,26 @@ export default function MovieCard({ movie }) {
 
     return (
         <Card className={classes.root} onClick={() => openMovieDetails()}>
+            <div className={classes.details}>
+                <CardContent className={classes.content}>
+                    <Typography component="h5" variant="h5">
+                        {movie.title}
+                    </Typography>
+                    <Typography variant="subtitle1" color="textSecondary">
+                        Release date : {movie.release_date}
+                    </Typography>
+                </CardContent>
+                <div className={classes.controls}>
+                    <IconButton aria-label="play/pause">
+                        <FavoriteBorder className={classes.favIcon} />
+                    </IconButton>
+                </div>
+            </div>
             <CardMedia
-                className={classes.media}
-                image={movie.imageUrl ? movie.imageUrl : ''}
-                title="Paella dish"
+                className={classes.cover}
+                image={`http://image.tmdb.org/t/p/w185/${movie.poster_path}`}
+                title={movie.title}
             />
-            <CardContent>
-                <Typography
-                    className={classes.title}
-                    color="textSecondary"
-                    gutterBottom
-                >
-                    {movie.director}
-                </Typography>
-                <Typography variant="h5" component="h2">
-                    {movie.title}
-                </Typography>
-                <Typography className={classes.pos} color="textSecondary">
-                    {movie.year}
-                </Typography>
-            </CardContent>
-            {/* <CardActions>
-                <Button size="small">More details</Button>
-            </CardActions> */}
         </Card>
     );
 }